requirements.txt
# --- Core AI / Speech Libraries ---
openai-whisper==20240930
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13

# --- FastAPI + Web Server ---
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3

# --- Utilities ---
numpy>=1.26.0
torch>=2.1.0
json5==0.9.25

# --- Optional helpers ---
pipwin==0.5.2
